Extract and clean up HTML Writer
