+++ src/main/java/org/opentripplanner/graph_builder/issue/report/DataImportIssuesToHTML.java
@@ -5,17 +5,13 @@ import com.google.common.collect.ImmutableSortedMultiset;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Multiset;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import org.opentripplanner.datastore.api.CompositeDataSource;
-import org.opentripplanner.datastore.api.DataSource;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssue;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueStore;
 import org.opentripplanner.graph_builder.model.GraphBuilderModule;
@@ -76,7 +72,7 @@ public class DataImportIssuesToHTML implements GraphBuilderModule {
         .stream()
         .collect(Collectors.groupingBy(DataImportIssue::getType));
 
-      // Sort each issue type by priority and convert issues to HTML string
+      // Sort each issue type by priority
       for (Map.Entry<String, List<DataImportIssue>> entry : sortedIssuesByType.entrySet()) {
         addIssues(
           entry.getKey(),
@@ -108,10 +104,10 @@ public class DataImportIssuesToHTML implements GraphBuilderModule {
       }
 
       try {
-        HTMLWriter indexFileWriter = new HTMLWriter("index");
+        HTMLWriter indexFileWriter = new HTMLWriter(reportDirectory, "index");
         indexFileWriter.writeFile(sortedIssueTypes);
       } catch (Exception e) {
-        LOG.error("Index file coudn't be created:{}", e.getMessage());
+        LOG.error("Index file couldn't be created:{}", e.getMessage());
       }
 
       LOG.info("Data import issue logs are in {}", reportDirectory.path());
@@ -129,7 +125,7 @@ public class DataImportIssuesToHTML implements GraphBuilderModule {
   public void checkInputs() {}
 
   /**
-   * Delete report if it exist, and return true if successful. Return {@code false} if the {@code
+   * Delete report if it exists, and return true if successful. Return {@code false} if the {@code
    * reportDirectory} is {@code null} or the directory can NOT be deleted.
    */
   private boolean deleteReportDirectoryAndContent() {
@@ -173,13 +169,13 @@ public class DataImportIssuesToHTML implements GraphBuilderModule {
       for (List<DataImportIssue> partition : partitions) {
         issueTypeOccurrences.add(issueTypeName);
         int labelCount = issueTypeOccurrences.count(issueTypeName);
-        file_writer = new HTMLWriter(issueTypeName + labelCount, partition);
+        file_writer = new HTMLWriter(reportDirectory, issueTypeName + labelCount, partition);
         writers.add(file_writer);
       }
     } else {
       issueTypeOccurrences.add(issueTypeName);
       int labelCount = issueTypeOccurrences.count(issueTypeName);
-      file_writer = new HTMLWriter(issueTypeName + labelCount, issues);
+      file_writer = new HTMLWriter(reportDirectory, issueTypeName + labelCount, issues);
       writers.add(file_writer);
     }
   }
@@ -194,133 +190,6 @@ public class DataImportIssuesToHTML implements GraphBuilderModule {
         e.getLocalizedMessage(),
         e
       );
-    }
-  }
-
-  class HTMLWriter {
-
-    private final DataSource target;
-
-    private final Collection<DataImportIssue> writerIssues;
-
-    private final String issueTypeName;
-
-    HTMLWriter(String key, Collection<DataImportIssue> issues) {
-      LOG.debug("Making file: {}", key);
-      this.target = reportDirectory.entry(key + ".html");
-      this.writerIssues = issues;
-      this.issueTypeName = key;
-    }
-
-    HTMLWriter(String filename) {
-      LOG.debug("Making file: {}", filename);
-      this.target = reportDirectory.entry(filename + ".html");
-      this.writerIssues = null;
-      this.issueTypeName = filename;
-    }
-
-    private void writeFile(Iterable<Multiset.Entry<String>> classes) {
-      try (
-        PrintWriter out = new PrintWriter(target.asOutputStream(), true, StandardCharsets.UTF_8)
-      ) {
-        out.println("<html><head><title>Graph report for OTP Graph</title>");
-        out.println("\t<meta charset=\"utf-8\">");
-        out.println("<meta name='viewport' content='width=device-width, initial-scale=1'>");
-        out.println("<script src='http://code.jquery.com/jquery-1.11.1.js'></script>");
-        out.println(
-          "<link rel='stylesheet' href='http://yui.yahooapis.com/pure/0.5.0/pure-min.css'>"
-        );
-        String css =
-          "\t\t<style>\n" +
-          "\n" +
-          "\t\t\tbutton.pure-button {\n" +
-          "\t\t\t\tmargin:5px;\n" +
-          "\t\t\t}\n" +
-          "\n" +
-          "\t\t\tspan.pure-button {\n" +
-          "\t\t\t\tcursor:default;\n" +
-          "\t\t\t}\n" +
-          "\n" +
-          "\t\t\t.button-graphwide,\n" +
-          "\t\t\t.button-parkandrideunlinked,\n" +
-          "\t\t\t.button-graphconnectivity,\n" +
-          "\t\t\t.button-turnrestrictionbad\t{\n" +
-          "\t\t\t\tcolor:white;\n" +
-          "\t\t\t\ttext-shadow: 0 1px 1px rgba(0, 0, 0, 0.2);\n" +
-          "\t\t\t}\n" +
-          "\n" +
-          "\t\t\t.button-graphwide {\n" +
-          "\t\t\t\tbackground: rgb(28, 184, 65); /* this is a green */\n" +
-          "\t\t\t}\n" +
-          "\n" +
-          "\t\t\t.button-parkandrideunlinked {\n" +
-          "\t\t\t\tbackground: rgb(202, 60, 60); /* this is a maroon */\n" +
-          "\t\t\t}\n" +
-          "\n" +
-          "\t\t\t.button-graphconnectivity{\n" +
-          "\t\t\t\tbackground: rgb(223, 117, 20); /* this is an orange */\n" +
-          "\t\t\t}\n" +
-          "\n" +
-          "\t\t\t.button-turnrestrictionbad {\n" +
-          "\t\t\t\tbackground: rgb(66, 184, 221); /* this is a light blue */\n" +
-          "\t\t\t}\n" +
-          "\n" +
-          "\t\t</style>\n" +
-          "";
-        out.println(css);
-        out.println("</head><body>");
-        out.println(
-          String.format("<h1>OpenTripPlanner data import issue log for %s</h1>", issueTypeName)
-        );
-        out.println("<h2>Graph report for <em>graph.obj</em></h2>");
-        out.println("<p>");
-
-        //adds links to the other HTML files
-        for (Multiset.Entry<String> htmlIssueType : classes) {
-          String label_name = htmlIssueType.getElement();
-          String label;
-          int currentCount = 1;
-          //it needs to add link to every file even if they are split
-          while (currentCount <= htmlIssueType.getCount()) {
-            label = label_name + currentCount;
-            if (label.equals(issueTypeName)) {
-              out.printf(
-                "<button class='pure-button pure-button-disabled button-%s' style='background-color: %s;'>%s</button>%n",
-                label_name.toLowerCase(),
-                IssueColors.rgb(label_name),
-                label
-              );
-            } else {
-              out.printf(
-                "<a class='pure-button button-%s' href=\"%s.html\" style='background-color: %s;'>%s</a>%n",
-                label_name.toLowerCase(),
-                label,
-                IssueColors.rgb(label_name),
-                label
-              );
-            }
-            currentCount++;
-          }
-        }
-        out.println("</p>");
-        if (writerIssues != null) {
-          out.println("<ul id=\"log\">");
-          writeIssues(out);
-          out.println("</ul>");
-        }
-
-        out.println("</body></html>");
-      }
-    }
-
-    /**
-     * Writes issues as LI html elements
-     */
-    private void writeIssues(PrintWriter out) {
-      String FMT = "<li>%s</li>";
-      for (DataImportIssue it : writerIssues) {
-        out.printf(FMT, it.getHTMLMessage());
-      }
     }
   }
 }
